requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pymongo==4.7.2
redis==5.0.4
python-dotenv==1.0.1
pydantic==2.7.1
requests>=2.32.2
streamlit==1.35.0
streamlit-folium==0.20.0
folium==0.16.0
pandas==2.2.2
reportlab==4.2.0
websocket-client==1.8.0
scikit-learn==1.5.0
numpy==1.26.4
websockets>=12.0